# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TiOS API",
        "description": "Welcolme to TiOS Platfor API documentation.",
        "termsOfService": "",
        "contact": {
            "email": "luis@tiosplatform.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:5000/api",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Returns list of projects",
                "operationId": "getProjects",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/project/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project information",
                "description": "Returns project data",
                "operationId": "getProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/projects/count": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get count of projects",
                "description": "Returns count of projects",
                "operationId": "getTotalProjectsCount",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/project/{project_id}/organization": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get Organization information by Project ID",
                "description": "Returns Organization data by Project ID",
                "operationId": "getOrganizationByProjectId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/project/{project_id}/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users by Project ID",
                "description": "Returns list of users by Project ID",
                "operationId": "getProjectUsers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/project/{project_id}/users/count": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get Users count by Project",
                "description": "Returns Users count by project id",
                "operationId": "getProjectUsersCount",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/project/{project_id}/dashboards/count": {
            "get": {
                "tags": [
                    "Dashboards"
                ],
                "summary": "Get Dashboards count by Project",
                "description": "Returns Dashboards count by project id",
                "operationId": "getProjectDashboardsCount",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/project/{project_id}/dashboards": {
            "get": {
                "tags": [
                    "Dashboards"
                ],
                "summary": "Get Dashboards by Project ID",
                "description": "Returns Dashboards data by project id",
                "operationId": "getDashboardByProjectId",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "project id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "passport": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token",
            "scopes": {}
        }
    },
...
}